refactor(ModalMedia): move component folder
