+++ packages/core/src/components/Modal/ModalMedia/ModalMedia.module.scss
