+++ packages/core/src/components/Modal/ModalMedia/ModalMedia.types.ts
